fix on benchmark build & use size_t for stack size

+++ include/rapid_string.h
@@ -334,7 +334,7 @@ inline bool rs_empty(const rapid_string *s);
  * @param s An initialized stack string.
  * @returns The stack string length.
  */
-inline uint8_t rs_stack_size(const rapid_string *s);
+inline size_t rs_stack_size(const rapid_string *s);
 
 /**
  * Returns the length of a heap string.
@@ -471,7 +471,7 @@ inline void rs_steal_n(rapid_string *s, char *buffer, size_t cap);
  * @param s An initialized stack string.
  * @param n The new size.
  */
-inline void rs_stack_resize(rapid_string *s, uint8_t n);
+inline void rs_stack_resize(rapid_string *s, size_t n);
 
 /**
  * Resizes a heap string. The new size must be smaller than `s->heap.capacity`.
@@ -742,7 +742,7 @@ inline bool rs_empty(const rapid_string *s)
 	return rs_size(s) == 0;
 }
 
-inline uint8_t rs_stack_size(const rapid_string *s)
+inline size_t rs_stack_size(const rapid_string *s)
 {
 	RS_ASSERT_STACK(s);
 
@@ -817,7 +817,7 @@ inline void rs_stack_append_n(rapid_string *s, const char *input, size_t n)
 	RS_ASSERT_PTR(input);
 	RS_ASSERT(RS_STACK_CAPACITY >= rs_stack_size(s) + n);
 
-	const uint8_t stack_size = rs_stack_size(s);
+	const size_t stack_size = rs_stack_size(s);
 	memcpy(s->stack.buffer + stack_size, input, n);
 	rs_stack_resize(s, stack_size + n);
 }
@@ -882,7 +882,7 @@ inline void rs_steal_n(rapid_string *s, char *buffer, size_t n)
 	s->heap.capacity = n;
 }
 
-inline void rs_stack_resize(rapid_string *s, uint8_t n)
+inline void rs_stack_resize(rapid_string *s, size_t n)
 {
 	RS_ASSERT(RS_STACK_CAPACITY >= n);
 	
@@ -949,7 +949,7 @@ inline void rs_heap_init_g(rapid_string *s, size_t n)
 
 inline void rs_stack_to_heap(rapid_string *s, size_t n)
 {
-	const uint8_t stack_size = rs_stack_size(s);
+	const size_t stack_size = rs_stack_size(s);
 
 	char tmp[RS_STACK_CAPACITY];
 	memcpy(tmp, s->stack.buffer, stack_size);
